+++ src/Core/ContentManager/Permalink/PermalinkGenerator.php
@@ -226,7 +226,7 @@ class PermalinkGenerator implements PermalinkGeneratorInterface
         $preservePathTitle = $this->getPreservePathTitleAttribute($item);
 
         if ($preservePathTitle === true && isset($attributes['title_path']) === true) {
-            return $attributes['title_path'];
+            return urlencode($attributes['title_path']);
         }
 
         if (isset($attributes['title']) === true) {